Quest List Added
This is a javascript file that contains the list of quests inside of an
array. We can use this until we have these stored and pulled from a
database,. This file will be loaded by index.html with a <script> tag,
and the declared "quests" array will be recognized by any other piece of
javascript on the page.